Add liblzma-dev packaging
As for xz-utils, defer to liblzma0 for some basic documentation.

Since liblzma-dev does not depend on xz-utils, a symbolic link to
doc/xz-utils/faq.txt.gz would be inappropriate.  Include the FAQ
again verbatim instead.

Most of the README has nothing to do with liblzma-dev, so do not
include it.  Later section 1.3 “Documentation for liblzma” should
be copied to liblzma-dev.README.Debian, but for now we can do
without.