Link command line tools against static liblzma
On x86, using PIC code for xz-utils reportedly causes a 10%
performance hit on some tests.  The library is only 132k, so
sharing copies is not a very compelling reason to use dynamically
linked PIC code.  The library shares sources with the binaries in
question, so if it were to be rebuilt to fix an important bug, the
binaries would incidentally be rebuilt, too.

Using a non-PIC library or statically linking on x86 platforms only
(i386/k*) might also be worth a try, but using the static library
as upstream does is simpler.  As a bonus, this makes it easier to
copy xz or xzdec to another computer without liblzma installed.

After this change, xz-utils no longer depends on liblzma0.  Thus
it must include its own copies of basic documentation files rather
than link to files in /usr/share/doc/liblzma0.