Optimize xzdec, lzmadec for size
Following the advice in the PACKAGERS file, build xzdec with
options that make for a smaller binary.

Instead of building twice with a “make clean” between, build each
version in a different build directory.  This lets us continue to
use dh’s built-in build sequence and makes it easier to inspect
the result of the build.

In the future, perhaps the two builds could take place in parallel.

As a downside, now xzdec and lzmadec are not tested in the course
of the build.