debian/symbols: update upstream version numbers
Without this change, Lintian gets confused:

	W: xz-utils symbols-file-contains-debian-revision on symbol lzma_alone_decoder@Base and 85 others

The symbols refer to the old upstream version, but there is no way
for Lintian to know that: they contain a dash and they are not
listed in debian/changelog.

For the initial packaging, it is cleaner not to refer to older
versions, anyway.  So Lintian pointed out a real improvement to be
made, though the message does not make it clear.

This problem should not recur in later versions, since each
upstream version listed should be either an official releases
(with no dash in the name) or a snapshot listed in
debian/changelog.