xz-utils: Conflicts: lzip
Debian packages xz-utils and lzip both provide an “lzdiff” command.
If both can be used to compare .lzma files, then the right way to
fix this is the alternatives mechanism, but it is not clear from
the lzip manual that lzdiff can be used that way.  Make xz-utils
conflict with lzip for now to be safe.

Reported by Piotr Engelking at <http://bugs.debian.org/540232>.