Add new xz-lzma package
Separate LZMA Utils compatibility executables and symlinks into
their own binary package, so xz-utils can coexist with lzma if
they are not installed.

It would be nice to avoid giving xz-lzma its own documentation
directory (instead making /usr/share/doc/xz-lzma a symlink to
/usr/share/doc/xz-utils), but it is not obvious how to do this
with dh.

The manpages installed with xz-lzma are symbolic links to the
manpages for xz-utils.  In Debian the targets of these links are
compressed with gzip, whereas the upstream Makefile makes links to
the uncompressed versions.  We should use a Makefile fragment
in debian/rules to make the links to compressed versions from the
links to the uncompressed versions, but for now, it is easier to
list the required links by hand.