xz-lzma: install lzgrep and lzdiff after all, Conflicts: lzip
It turns out that the commands from Lzip cannot handle .lzma
files after all.  Providing lzdiff and so on for backward
compatibility requires conflicting with Lzip, for now.

That’s the bad news.  The good news is that Lzip drops those
commands in version 1.8-rc2.

This reverts commits 92cb95b (“xz-lzma description: Suggest lzip
for lzgrep, too”, 2009-08-09) and a9c0474 (“debian/rules: xz-lzma:
do not install lzdiff and lzgrep”, 2009-08-08) and makes the
Conflicts: field in debian/control more specific.