debian/changelog: more bugs fixed by new upstream version
The Debian package for xz-utils 4.999.8beta (see commit 8500893,
2009-06-27) is a bit broken, since the packaging is based on a
more recent upstream snapshot than the source files used.  In
particular, debian/copyright stated that the sources are in the
public domain but that version is not, and the documentation
referred to various files that were not part of the upstream doc/
directory yet.  Using more recent upstream sources fixes both
bugs.