debian/rules: move lzmainfo to xz-lzma package
Even though Debian lzma does not include it, lzmainfo is an LZMA
Utils compatibility command and so belongs in xz-lzma.  It is
tempting to include it in xz-utils so one can say “all the
functionality of xz-lzma is in xz-utils” without lying, but:

 * the small lzmadec binary is already only in xz-lzma; xzdec
   cannot decompress .lzma files

 * something like the functionality of lzmainfo should be
   available through “xz --list” once that is implemented

 * if Debian’s LZMA Utils packaging ever includes lzmainfo, it
   would be nice to avoid a conflict between xz-utils and lzma

 * other distros put lzmainfo in xz-lzma, so that is where many
   users will expect to find it

After this change, xz-utils does not need to depend on liblzma0,
so it installs its own documentation again.