xz-lzma: fix conflict with pseudo-essential package lzma
On switching from lzma to xz-lzma, the intended sequence of events
is as follows:

1. xz-utils is unpacked
2. libc6 is upgraded, if necessary
3. xz-lzma is unpacked, overwriting the files from lzma
4. lzma has no files left, so it is marked removed

Step 3 must occur after steps 1 and 2, or dpkg may find itself
without a functioning lzma command to use, so make xz-lzma
Pre-Depends: xz-utils, libc6 (>= 2.6).  During step 3, files
from xz-lzma and lzma coexist, so drop the Conflicts: lzma.

Ensuring that step 4 occurs is a little trickier.  Right now
xz-lzma provides all the files that lzma does, but that could
change in some later version of lzma.  If it does, on some systems,
lzma and xz-lzma would both appear to be installed, but there
should be no other problems.

Note that the xz binary which is the target of the /usr/bin/lzma
symbolic link is linked to the static liblzma, so liblzma0 is not
required for its essential functionality.