Update packaging for new upstream release
The content of bugs.txt has been moved to README, so do not try
to install bugs.txt any more.  Maybe later it would make sense to
add reportbug input, but for now, users of liblzma0 wanting
debugging help can ask or look in the xz-utils documentation
directory.

No API changes, but bump the version numbers in the symbols file
anyway.  The symbols file does not need to distinguish versions
not packaged for Debian.

Note: the .diff.gz for this version is not just the diff from the
v4.999.9beta tag to this commit.  It also adds files
debian/changelog.upstream (generated with
debian/changelog.upstream.sh) and {m4,po}/.gitignore (from
v4.999.9beta but not included in upstream’s release tarball).
Hopefully, this procedure will become simpler in future releases.