Drop xz-lzma again
It looks like using alternatives for the lzma command is not the
right thing after all.  Until dpkg can be installed without lzma,
the lzma command has to be usable at all times, but the
alternatives system does not provide a usable lzma command until
the package is configured.  This was not causing trouble with
the existing package only because no one tried unpacking an
lzma-compressed .deb with xz-lzma unconfigured.

Once dpkg can be installed without lzma, the lzma command will
no longer be essential, and we can add back an xz-lzma that
conflicts with lzmp.

This reverts commit 578df4d0936eb5bc6770be60f57ac6cd295b654b.

This also moves the lzmainfo and lzmadec commands to the xz-utils
package, since they are not available from elsewhere, and adds
some documentation.