debian/rules: Avoid delayed string processing
Use := instead of = to define variables when the right-hand side
requires some processing.  This follows the principle of least
surprise and should also speed up the build by some small amount.

In portable makefiles, one would avoid := as a feature specific to
GNU make, but this makefile would be hard to write portably for
other reasons anyway.