changelog.upstream.sh: cope with Debian revision -2
The code in debian/changelog.upstream.sh was written with the
naïve assumption that each Debian revision should have a
changelog entry of the form

	* New upstream snapshot, taken from upstream commit <blah>.

or

	* New upstream release.

But sometimes a new Debian revision appears between merges from
upstream.  Handle this case by checking when the upstream
component of the version number changes.