debian/rules: Work around build failure on the Hurd
The assembler code in src/liblzma/check uses __MACH__ to detect
Mac OS X, but __MACH__ is set on Hurd, too, which is a pretty
ordinary ELF-using system.

This is trivially fixed by checking for __APPLE__ as well, but
rather than patching the upstream sources, it is safer to disable
the assembler code on Hurd for now.