debian/copyright: Remove big list of copyright holders
The list of copyright notices can already be extracted with
"grep Copyright /usr/share/doc/xz-utils/copyright" or
"grep © /usr/share/doc/xz-utils/copyright".

It requires worthless effort to maintain an extra copy of the
same list at the top of the copyright file.  It is even a little
misleading, since these copyright statements apply to individual
files, not the work as a whole.