debian/rules: allow simple globs in po/.gitignore
Add a perl script to interpret simple globs and comments in
po/.gitignore.  It is intentionally very restrictive: filenames are
only allowed to contain alphanumeric characters, dashes, at-signs, and
periods.

The only supported glob is a * at the beginning of a filename; this
restriction is posed not for ease of implementation but to avoid
pretending to support full filename globbing.  We only want globs to
match the current directory, but .gitignore syntax allows matching
subdirectories as well, so there has to be some awkwardness here.