debian/rules: use --enable-assembler on Hurd
Start using the optimized CRC32 implementation on the Hurd.  It would
be more comfortable to wait until it is enabled by default in
configure.ac, but that would create a chicken-and-egg problem, since
it would also be more comfortable for upstream to wait until the code
has been seen deployed without trouble.

The --enable-assembler option requires a parameter specifying the CPU
type to work (x86 or x86_64).