New xz-utils-tiny package for tiny decompressors
These tiny decompressors are mostly useful because they are small, for
copying to another machine.  Unfortunately, they greatly increase the
size of the standard xz-utils package which already provides all their
functionality.  So split the extra tools off into a separate package.

Maybe it would make sense to shrink that package by giving it
dedicated a copyright and changelog file that only describes xzdec and
lzmadec themselves.  But that package is not essential and standard
Debian is not very useful on incredibly tiny machines anyway, so this
can wait.

Reported-by: Guillem Jover <guillem@debian.org>
Fixes: http://bugs.debian.org/571060