Rename xz-utils-tiny to xzdec
Packages with names of the form foo-tiny tend to represent drop-in
replacements for a large package with just the piece everyone needs.
(Examples: vim-tiny, mlterm-tiny.)  The packaging for xzdec and
lzmadec could be modified to provide the same thing, but it is not
obviously necessary, and in any event, the current packaging does not
provide that.

About the name xzdec:  True, the package provides lzmadec, too, but
features like this are what the long description can describe.  This
name at least appropriately suggests what the package provides: an XZ
decompressor invoked under a different name.