Do not needlessly install and run doxygen on buildds
Build systems for slow architectures (like mips) can save some time by
avoiding building architecture-independent packages such as
liblzma-doc.

Due to an sbuild bug, Build-Depends-Indep might not be satisfied when
running the build target.  We have to check the build-dependency
ourselves at run time to detect that case (in which case this is an
architecture-specific build so we can skip the doxygen run).  This
way, buildds that run the build target (all of them, with current
sbuild) will not waste their time building documentation unless
doxygen happens to be present already.