debian/control: use the new autopoint package
The autopoint script is going to move in Debian from the gettext
package to a new package of its own.  Adjust.

It might be easier on backporters to use autopoint | cvs, but it
is would be easy for them to make this change themselves, so I do
not think it is worth it.