debian/copyright: remove overall license notice
In some (pretty rare) circumstances, it might be useful to list one
license document such that following its terms is enough to respect
all copyright in a package.  This would be nice to have because it
allows a user to avoid worrying about details, but on the other hand,
any license document meeting these requirements is going to have
fairly strict terms.

This package is no exception.  Most of the package has been
released into the public domain, but:

 - The comments in Doxyfile.in are under the GPL-2, as far as I
   can tell.  Of course, they do not make their way into the
   binary package;

 - The autoconf macros in m4/acx_pthread.m4 are under the GPL-3+,
   with exceptions to the terms to avoid greatly restricting the
   freedom to use, modify, and distribute the generated
   configure script and binary package.

So just remove the misleading notice about overall license.  If
there is heavy demand, some more precise per-binary-package
copyright files can be added later.