Add back xz-lzma package
Now that dpkg no longer depends on lzma, it is possible to ship
a package that conflicts with lzma.  The new xz-lzma is an alternative
to the lzma package, providing the same commands with an
implementation from XZ Utils.

This package conflicts with lzma rather than using alternatives
to follow the principle of least surprise.  The main downside to
this approach is that a multi-user installation with both
implementations available and xz-lzma as default is not possible
without custom symlinks in /usr/local/bin.  This can be revisited later.

The lzmainfo command is in xz-utils, not moved to the new package.
Until the xz command learns an “xz --format=lzma --list” mode, without
that command xz-utils would be functionally incomplete.

Most files in the documentation directory are symlinks to save space.