debian: Simplify upgrade path from ancient xz-lzma and lzip
Use Replaces instead of Conflicts to represent conflicts in
debian/control, as recommended in policy §7.4.

This should make it easier for package managers to come up with
an upgrade path: it is no longer necessary to remove or upgrade
old xz-lzma in order to upgrade xz-utils.

Use Breaks to force the obsolete packages not to linger for long
so scenarios in which the replacing package is removed and breaks
them become less likely.