changelog.upstream.sh: cope with Debian revision -2, again
debian/changelog.upstream.sh was written with the naïve assumption
that each non-snapshot Debian revision should correspond to a
different upstream release.

But sometimes a new Debian revision is needed before it is time to
merge from upstream again.  Check for this case explicitly to
avoid unsightly

	Version 5.0.0; changes since 5.0.0:
	-----------------------------------

entries in the upstream changelog.