debian/rules: nocheck support
Allow suppression of built-time tests via the
DEB_BUILD_OPTIONS=nocheck option.

The effect on time to build is not very dramatic since the current
xz-utils test suite is very small.  This is most useful when tests are
failing on some platform, since it provides an easy way to finish the
build anyway.

See policy §4.9.1 for details.

Reported-by: Thorsten Glaser <tg@mirbsd.de>