Install liblzma2 under /usr/lib/<triplet>/ instead of /usr/lib/
This change has been cooking in experimental since June without
problems.  Time to add it to sid and get some bug reports.

This patch doesn’t add the Multi-Arch fields to debian/control yet
because they are not documented in policy and require dpkg support
(added in version 1.15.4).  But since that version is already in
stable, the Multi-Arch fields will probably follow soon.