debian/rules: split build-arch, build-indep targets
In the not-so-distant future, Debian's autobuilders will start using
the build-arch and build-indep targets instead of unconditionally
invoking “debian/rules build”.  That will be nice --- in particular,
it would allow the hack introduced in commit 8f68eeb (Do not
needlessly install and run doxygen on buildds, 2010-02-23) to be
removed.  At the moment, we can prepare by making sure those targets
are provided and do just the right thing.

Until recently, that was not easy --- the simplest way involved doing
something like the following:

	override_dh_auto_build:
		set -e; \
		if \
			dh_listpackages | \
			grep liblzma-doc >/dev/null; \
		then \
			doxygen Doxyfile; \
		fi
		dh_auto_build

That is because debhelper’s sequences do not include separate items
for building the architecture-dependent and architecture-independent
part of a package --- internally, dh wants to treat “dh build-arch”
somewhat like “dh build -a”.

Luckily debhelper 8.9.0 gives a way out: we can define build-arch,
build-indep, and the targets they depend on as ordinary makefile
targets in debian/rules and dh will invoke them at the appropriate
moment in sequences like “dh build” and “dh binary-arch”.  Use it.

To avoid running “configure” twice, use the makefile output by
configure as a stamp file.  To avoid running the test suite twice, put
a stamp file at debian/test-stamp.  The former could be improved by
listing the appropriate dependencies (to allow bisecting and other
workflows that involving tweaking makefiles and incrementally
rebuilding), while the latter is unavoidable.  Both were previously
taken care of by *.debhelper.log.