Provide LZMA Utils compatibility commands through the alternatives system
This way, there is no need for the tiny xz-lzma package whose
installation or non-installation reflects whether the sysadmin wanted
XZ Utils to provide commands such as lzma and unlzma.

Something like this seemed like a good idea and was then backed out as
long ago as cf1aca1 (Drop xz-lzma again, 2009-09-30).  Back then, lzma
was pseudo-essential due to dpkg's dependency on it, making the
temporary removal of /usr/bin/lzma during an upgrade of the "lzma"
package in preparation for providing it through the alternatives
system awkward.

That is not a problem any more:

 - squeeze dpkg already does not use the "lzma" command.  No problem
   in upgrades from squeeze.

 - etch-m68k dpkg suggests lzma as an experimental feature and no
   Debian package uses LZMA compression for its files yet.  No danger
   there, either.

 - people trying to upgrade using lenny dpkg are doing something
   unsupported and are on their own.  Luckily, it shouldn't be a
   problem for them, either --- once the new "lzma" is extracted,
   configuring it is sufficient to make the lzma command available
   again.

After this change, you can allow xz to implement the "lzma" command
with a simple "update-alternatives --config lzma".  No more need to
fish for a Debian installation CD or network cable.

The update-alternatives priority is 20 (less than the a neutral 30)
so the lzma package gets to provide those commands by default when
installed.