Match upstream ABI at last
It turns out that due to upstream's careful planning it is pretty safe
to have a process image containing both liblzma2 and liblzma5, since
the changes in struct layout are only in trailing fields that no
exported functions need to touch (and no one seems to rely on the size
of these structs).  So let's match upstream and avoid some trouble.

This simplifies a number of pieces and allows us to get rid of the
documentation in liblzma’s and xz-utils’s READMEs.Debian on
binary-incompatibility with upstream.