liblzma: define a base version of lzma_code for ABI-compatibility
Generally speaking, it should be safe to have a process image
including both the pre-5.0 and v5.0 implementations of liblzma
(liblzma2 and liblzma5 in Debian), since the only changes in ABI were
to remove a function and to rearrange reserved fields which are not
used by most functions.

There is one function that does differ between the two ABIs, though:
lzma_code() checks that the reserved fields at the end of an
lzma_stream are all 0 before proceeding.  Applications linked directly
to liblzma2 and indirectly to liblzma5 use the implementation from
liblzma5, which looks at more reserved fields than the application
knows to fill, resulting in segfaults or errors:

	$ LD_PRELOAD=/usr/lib/x86_64-linux-gnu/liblzma.so.5 xz README
	xz: README: Unsupported options

Unfortunately one cannot even get lucky and find the symbol from
liblzma2 used from time to time: versioned symbols take precedence
over unversioned ones when resolving unversioned references.

Fortunately the fix is clear: if we additionally bind a compatibility
version of lzma_code() to the unspecified base version of lzma_code,
_it_ will be used to resolve references from code linked to liblzma2
when liblzma2 and liblzma5 share a process image.  Do so.

Caveat: libbfd-based binutils handles this well (and it is documented
to work), but GNU gold 2.21.90.20111004 errors out with

	/usr/bin/ld: error: symbol lzma_code has undefined version

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>