debian: correct lzma_code@Base rationale
debian/patches/abi-liblzma2-compat says:

| Applications linked directly to liblzma2 and indirectly to liblzma5 use
| the implementation from liblzma5

As Jakub Wilk explained, that doesn't seem to be the case.  If a
program is linked to two versions of a library, one of which doesn't
use versioned symbols, then the symbols from the directly-linked one
shadow the indirectly-linked one.

Annoyingly, I noticed the same thing last year and then promptly
forgot about it.  Luckily in cases where liblzma2 wins the right to
provide lzma_code(), that just means the lib checks a few fewer
reserved fields and the patch is not needed, and in cases where
liblzma5 wins, the patch works.

Reported-by: Jakub Wilk <jwilk@debian.org>