Do not let liblzma.a from xzdec-build clobber the real thing
After commit 18be97e6 (Optimize xzdec, lzmadec for size, 2009-06-24),
the install target in debian/rules takes place in two stages: first
install tools from the normal build, then from the xzdec build.  In
particular, after installing a perfectly good liblzma.a, we overwrite
it with the copy from the xzdec build with all encoders disabled ---
oops.

The fix is simple: let the install rule for xzdec run _before_ the
normal build.

Probably this wasn't noticed before because static linking on Debian
is not popular.  The xz and lzmainfo commands, shared library, and
scripts are all disabled when building xzdec, so this problem only
affects the static library.

Reported-by: Anton Tolchanov <anton.tolchanov@gmail.com>
Fixes: http://bugs.debian.org/673001