debian/copyright: remove unwanted periods in formatted form
In Debian's machine-readable copyright format, the space between
paragraphs is indicated using a dot (.) preceded by a single space.
Some instances of this were changed to a dot preceded by two spaces
(which signifies a literal '.') by mistake when marking the paragraphs
around them as literal to avoid automatically rewrapping.