xz-utils/README.Debian: Document patches
End-users and distributors might wonder how a stable release is made
out of an alpha release and how the packaged code differs from the
original code.

Briefly describe each patch in README.Debian to help such people,
imitating the gcc packaging.  This information does not belong in the
package description because it is more relevant when using the package
than when deciding whether to install it.  The package description
already mentions that threads are not supported.

Thanks to Andrew Pollock, Thorsten Glaser, and Lasse Collin for
advice.