debian/control: Remove build-dependency on perl
xz-utils' build-dependency on perl is unnecessary, and it impedes
cross-building (because perl is assumed to be needed from the host
architecture by default), which is awkward because xz-utils is part of a
base buildd chroot.  perl:native would be better, but Debian's build
daemons don't support that syntax yet.

As it happens, the perl build-dependency is entirely unnecessary.
xz-utils only uses the bare perl interpreter, without modules (other
than 'warnings' and 'strict').  As such, it has no need for
non-Essential facilities, and can rely on the Essential perl-base.  In
private mail, Jonathan suggested rewriting in awk, but this is
unnecessary; you really should just be able to rely on the Essentialness
of the perl interpreter here, and the Perl policy sanctions this given
that xz-utils is part of the base system.

  * Remove build-dependency on perl; we use no non-trivial Perl modules so
    the Essential perl-base is perfectly adequate, and an unadorned
    build-dependency on perl is problematic for cross-building.

[helmut@subdivi.de: uploaded as NMU]