kfreebsd: link against libfreebsd-glue
cpuset_getaffinity() is available on kfreebsd but not in libc but in
libfreebsd-glue. This code adds the library if kfreebsd is detected so
cpuset_getaffinity() can be used.
Without it tuklib_cpucores will detect `cpuset' because the compile test
succeeds and later it fails at link phase.

Patch-Name: kfreebsd-link-against-libfreebsd-glue.patch
Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>